+++ backend/canisters/community/api/src/updates/c2c_bot_send_message.rs
@@ -1,6 +1,6 @@
 use crate::send_message;
 use serde::{Deserialize, Serialize};
-use types::{BotInitiator, BotMessageContent, ChannelId, MessageId, MessageIndex, UserId};
+use types::{BotInitiator, BotMessageContent, ChannelId, EventIndex, GroupReplyContext, MessageId, MessageIndex, UserId};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -9,6 +9,7 @@ pub struct Args {
     pub channel_id: ChannelId,
     pub thread_root_message_index: Option<MessageIndex>,
     pub message_id: MessageId,
+    pub replies_to: Option<EventIndex>,
     pub content: BotMessageContent,
     pub bot_name: String,
     pub block_level_markdown: bool,
@@ -24,7 +25,7 @@ impl From<Args> for send_message::Args {
             content: value.content.into(),
             sender_name: value.bot_name,
             sender_display_name: None,
-            replies_to: None,
+            replies_to: value.replies_to.map(|r| GroupReplyContext { event_index: r }),
             mentioned: vec![],
             forwarding: false,
             block_level_markdown: value.block_level_markdown,