+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove bot API Keys ([#8131](https://github.com/open-chat-labs/open-chat/pull/8131))
 - Remove _v2 suffix from bot_\* endpoints ([#8139](https://github.com/open-chat-labs/open-chat/pull/8139))
 - Support bot making inline replies ([#8146](https://github.com/open-chat-labs/open-chat/pull/8146))
+- Return newly created channel Ids after creating community ([#8156](https://github.com/open-chat-labs/open-chat/pull/8156))
 
 ### Fixed